src/VumbaSoft.AdventureWorks.Controllers/Sales/SpecialOffers/SpecialOffersController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Sales
{
    [Area("Sales")]
    public class SpecialOffersController : ValidatedController<ISpecialOfferValidator, ISpecialOfferService>
    {
        public SpecialOffersController(ISpecialOfferValidator validator, ISpecialOfferService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(SpecialOfferView offer)
        {
            if (!Validator.CanCreate(offer))
                return View(offer);

            Service.Create(offer);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<SpecialOfferView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<SpecialOfferView>(id));
        }

        [HttpPost]
        public ActionResult Edit(SpecialOfferView offer)
        {
            if (!Validator.CanEdit(offer))
                return View(offer);

            Service.Edit(offer);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<SpecialOfferView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
